+++ gdb/samples/define.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 make hello
-gdb --quiet ./hello < define.gdb
+gdb --quiet ./hello -x define.gdb