+++ gdb/samples/disassemble.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 make hello
-gdb --quiet ./hello < disassemble.gdb
+gdb --quiet ./hello -x disassemble.gdb