+++ gdb/samples/global-var.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 make global-var
-gdb -q ./global-var < global-var.gdb
+gdb -q ./global-var -x global-var.gdb