+++ gdb/samples/info-address.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 make hello
-gdb --quiet ./hello < info-address.gdb | grep -ow -E '0x[0-9a-fA-F]*?'
+gdb --quiet ./hello -x info-address.gdb | grep -ow -E '0x[0-9a-fA-F]*?'