+++ gdb/samples/set-value.sh
@@ -2,6 +2,6 @@
 
 make set-value set-value-func set-value-volatile
 
-gdb -q set-value < set-value.gdb
-gdb -q set-value-func < set-value.gdb
-gdb -q set-value-volatile < set-value.gdb
+gdb -q set-value -x set-value.gdb
+gdb -q set-value-func -x set-value.gdb
+gdb -q set-value-volatile -x set-value.gdb