+++ gdb/samples/symbol-file.sh
@@ -1,3 +1,4 @@
 #!/bin/bash
 make hello.strip
-gdb --quiet hello.strip < symbol-file.gdb
+# Skip: Enable debuginfod for this session? (y or [n])
+DEBUGINFOD_URLS="" gdb --quiet hello.strip -x symbol-file.gdb