+++ gdb/samples/until.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 make hello
-gdb --quiet ./hello < until.gdb
+gdb --quiet ./hello -x until.gdb