+++ gdb/samples/mmap-file.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+make hello
+gdb --quiet ./hello -x mmap-file.gdb